requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Database & Supabase
supabase==2.3.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI/ML - Embeddings
sentence-transformers==2.2.2
torch==2.2.0

# Validation & Data Models
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.24.1

# AWS SDK (for Secrets Manager)
boto3==1.34.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
